TCK CircuitBreakerTest - test methods reusing CircuitBreakerClientDefaultSuccessThreshold Refactor
https://github.com/eclipse/microprofile-fault-tolerance/issues/323

The tests testCircuitInitialSuccessDefaultSuccessThreshold and
testCircuitLateSuccessDefaultSuccessThreshold were moved to an independent
test to avoid dependency between tests that uses the same bean in this
case CircuitBreakerClientDefaultSuccessThreshold, that will can generate
possible failures when the circuit breaker leaves open.
To ensure that each of these tests use a fresh circuit breaker were created
an independent deployment for this purpose.
