+++ tck/src/main/java/org/eclipse/microprofile/fault/tolerance/tck/CircuitBreakerTest.java
@@ -65,9 +65,8 @@ public class CircuitBreakerTest extends Arquillian {
                         .addAsManifestResource(EmptyAsset.INSTANCE, "beans.xml")
                         .as(JavaArchive.class);
 
-        WebArchive war = ShrinkWrap.create(WebArchive.class, "ftCircuitBreaker.war")
+        return ShrinkWrap.create(WebArchive.class, "ftCircuitBreaker.war")
                         .addAsLibrary(testJar);
-        return war;
     }
 
     /**
@@ -252,149 +251,6 @@ public class CircuitBreakerTest extends Arquillian {
         Assert.assertEquals(serviceAExecutions, 9, "The number of serviceA executions should be 9");
     }
 
-    /**
-     * Analagous to testCircuitDefaultSuccessThreshold but with a different success/failure pattern
-     * for the service that is called. In this case, the service initially succeeds.
-     * 
-     * With requestVolumeThreshold = 4, failureRatio=0.75 and successThreshold = 1 
-     * the expected behaviour is,
-     *
-     * Execution Behaviour 
-     * ========= ========= 
-     * 1         SUCCESS 
-     * 2         RunTimeException 
-     * 3         RunTimeException  
-     * 4         RunTimeException
-     * 5         CircuitBreakerOpenException
-     * Pause for longer than CircuitBreaker delay, so that it transitions to half-open
-     * 6         SUCCEED (CircuitBreaker will be re-closed as successThreshold is 1. The impact of
-     *                    the success of the service and the closure of the Circuit is to reset the
-     *                    rolling failure window to an empty state. Therefore another 4 requests need
-     *                    to be made - of which at least 3 need to fail - for the Circuit to open again)
-     * 7         SUCCESS
-     * 8         RunTimeException
-     * 9         RunTimeException
-     * 10        RuntimeException
-     * 11        CircuitBreakerOpenException
-     *
-     */
-    @Test
-    public void testCircuitInitialSuccessDefaultSuccessThreshold() {
-
-        // Reset the counter in serviceA, to ensure that the test's environment is reinitialized.
-        clientForCBDefaultSuccess.setCounterForInvokingServiceA(0);
-
-        for (int i = 1; i < 12; i++) {
-            int[] successSet = new int[]{1,5,6};
-            try {
-                clientForCBDefaultSuccess.serviceA(successSet);
-                if (!contains(new int[]{1, 6, 7}, i)) {
-                    Assert.fail("serviceA should throw an Exception in testCircuitDefaultSuccessThreshold on iteration " + i);
-                }
-            }
-            catch (CircuitBreakerOpenException cboe) {
-                // Expected on execution 5 & 11
-                if (!contains(new int[]{5, 11}, i)) {
-                    Assert.fail("in serviceA no CircuitBreakerOpenException should be fired on iteration " + i);
-                }
-                else if (i == 5) {
-                    // Pause to allow the circuit breaker to half-open
-                    try {
-                        Thread.sleep(2000);
-                    }
-                    catch (InterruptedException e) {
-                        e.printStackTrace();
-                    }
-                }
-            }
-            catch (RuntimeException ex) {
-                // Expected
-                if (!contains(new int[]{2, 3, 4, 8 , 9, 10}, i)) {
-                    Assert.fail("serviceA should not throw a RuntimeException on iteration " + i);
-                }
-            }
-            catch (Exception ex) {
-                // Not Expected
-                Assert.fail("serviceA should throw a RuntimeException or CircuitBreakerOpenException in testCircuitDefaultSuccessThreshold "
-                                + "on iteration " + i);
-            }
-        }
-        int serviceAExecutions = clientForCBDefaultSuccess.getCounterForInvokingServiceA();
-        Assert.assertEquals(serviceAExecutions, 9, "The number of serviceA executions should be 9");
-    }
-
-    /**
-     * Analagous to testCircuitDefaultSuccessThreshold but with a different success/failure pattern
-     * for the service that is called. In this case, the service succeeds in the last call in the 
-     * rolling window but the circuit will still open as the failureRatio has been breached.
-     * 
-     * With requestVolumeThreshold = 4, failureRatio=0.75 and successThreshold = 1 
-     * the expected behaviour is,
-     *
-     * Execution Behaviour 
-     * ========= ========= 
-     * 1         RunTimeException 
-     * 2         RunTimeException 
-     * 3         RunTimeException  
-     * 4         SUCCESS
-     * 5         CircuitBreakerOpenException
-     * Pause for longer than CircuitBreaker delay, so that it transitions to half-open
-     * 6         SUCCEED (CircuitBreaker will be re-closed as successThreshold is 1. The impact of
-     *                    the success of the service and the closure of the Circuit is to reset the
-     *                    rolling failure window to an empty state. Therefore another 4 requests need
-     *                    to be made - of which at least 3 need to fail - for the Circuit to open again)
-     * 7         RuntimeException
-     * 8         RunTimeException
-     * 9         RunTimeException
-     * 10        SUCCESS
-     * 11        CircuitBreakerOpenException
-     *
-     */
-    @Test
-    public void testCircuitLateSuccessDefaultSuccessThreshold() {
-
-        // Reset the counter in serviceA, to ensure that the test's environment is reinitialized.
-        clientForCBDefaultSuccess.setCounterForInvokingServiceA(0);
-
-        for (int i = 1; i < 12; i++) {
-            int[] successSet = new int[]{4,5,9};
-            try {
-                clientForCBDefaultSuccess.serviceA(successSet);
-                if (!contains(new int[]{4, 6, 10}, i)) {
-                    Assert.fail("serviceA should throw an Exception in testCircuitLateSuccessDefaultSuccessThreshold on iteration " + i);
-                }
-            }
-            catch (CircuitBreakerOpenException cboe) {
-                // Expected on execution 5 & 11
-                if (!contains(new int[]{5, 11}, i)) {
-                    Assert.fail("in serviceA no CircuitBreakerOpenException should be fired on iteration " + i);
-                }
-                else if (i == 5) {
-                    // Pause to allow the circuit breaker to half-open
-                    try {
-                        Thread.sleep(2000);
-                    }
-                    catch (InterruptedException e) {
-                        e.printStackTrace();
-                    }
-                }
-            }
-            catch (RuntimeException ex) {
-                // Expected
-                if (!contains(new int[]{1, 2, 3, 7, 8 , 9}, i)) {
-                    Assert.fail("serviceA should not throw a RuntimeException on iteration " + i);
-                }
-            }
-            catch (Exception ex) {
-                // Not Expected
-                Assert.fail("serviceA should throw a RuntimeException or CircuitBreakerOpenException in testCircuitDefaultSuccessThreshold "
-                                + "on iteration " + i);
-            }
-        }
-        int serviceAExecutions = clientForCBDefaultSuccess.getCounterForInvokingServiceA();
-        Assert.assertEquals(serviceAExecutions, 9, "The number of serviceA executions should be 9");
-    }
-    
     /**
      * A test to exercise Circuit Breaker thresholds, with a default
      * SuccessThreshold
